Develop (#27)
* feat: 회원가입 중복 로직 추가

* GitHub Actions 워크플로우 파일 생성

* GitHub Actions 워크플로우 파일
PR 액션 포함

* fix: 오류 수정

* appleboy action 버전 수정

* uses 버전 변경

* swagger 추가

* appleboy 버전 master로 변경

* 메이븐 빌드 과정 추가.
주석 변경

* 자바 21 명시

* logout 주석처리

* 빌드 시 데이터베이스 환경변수 추가

* db 정보 env로 수정

* 중괄호 수정

* 물품 수정 시 필드 각각 수정 가능하도록 update DTO,service 수정

* 빌드 성공한 jar 파일 위치 수정

* 도커 로그인 추가 및 주석 수정

* 컨테이너 실행 부분 수정

* 도커 부분에 sudo 추가

* 도커허브 이용하도록 수정

* NOT FOUND 예외 처리 구현

* swagger에 error message 반영

---------

+++ src/main/java/com/konkukrent/demo/service/ProductService.java
@@ -4,6 +4,7 @@ import com.konkukrent.demo.dto.ProductDto.ProductCreateRequestDto;
 import com.konkukrent.demo.dto.ProductDto.ProductResponseDto;
 import com.konkukrent.demo.dto.ProductDto.ProductUpdateRequestDto;
 import com.konkukrent.demo.entity.Product;
+import com.konkukrent.demo.exception.exceptionClass.CustomException;
 import com.konkukrent.demo.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -11,6 +12,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
+
+import static com.konkukrent.demo.exception.properties.ErrorCode.PRODUCT_NOT_FOUND;
 
 @Service
 @RequiredArgsConstructor
@@ -43,10 +46,21 @@ public class ProductService {
     public ProductResponseDto updateProduct(Long productId, ProductUpdateRequestDto request) {
         Optional<Product> optionalProduct = productRepository.findById(productId);
         if (optionalProduct.isEmpty()) {
-            throw new RuntimeException("Product not found");
+            throw new CustomException(PRODUCT_NOT_FOUND);
         }
         Product product = optionalProduct.get();
-        product.setRemainNumber(request.getRemainNumber());
+        if (request.getName() != null) {
+            product.setName(request.getName());
+        }
+        if (request.getTotalNumber() != null) {
+            product.setTotalNumber(request.getTotalNumber());
+        }
+        if (request.getRemainNumber() != null) {
+            product.setRemainNumber(request.getRemainNumber());
+        }
+        if (request.getRentalPeriod() != null) {
+            product.setRentalPeriod(request.getRentalPeriod());
+        }
         Product updatedProduct = productRepository.save(product);
         return mapToResponseDto(updatedProduct);
     }
@@ -54,6 +68,10 @@ public class ProductService {
     // 물품 삭제
     @Transactional
     public void deleteProduct(Long productId) {
+        Optional<Product> optionalProduct = productRepository.findById(productId);
+        if (optionalProduct.isEmpty()) {
+            throw new CustomException(PRODUCT_NOT_FOUND);
+        }
         productRepository.deleteById(productId);
     }
